requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.7
Flask==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-engineio==4.10.1
python-socketio==5.11.3
simple-websocket==1.1.0
uvicorn==0.18.3
Werkzeug==3.1.3
wsproto==1.2.0
